Fix bug when ending copy with non-aligned halves.
Add option to autoflag the DC channel in an attempt
to improve the dynamic range of the spectra (thanks
to the short-int encoding used).

Add option to flag the bad-C-board channels for the
same reason. This is genuinely necessary on the 
baselines between two bad-C-board antennas in XX; 
the quantization caused by the huge amplitudes causes
seriously data corruption otherwise.

Don't test for exact equality between times but rather
agreement within a tolerance. Apparently sometimes
they differ by ~1e-9.

A few output tweaks.



git-svn-id: https://svn.hcro.org/mmm@55 9b194dde-2b4d-0410-98f0-bba1668eb47e