Import the improved 'hex' observing scripts. These are not in any particular user's MMM subdirectory to emphasize their joint ownership by the MMM group as a whole.
The file headers have some basic documentation but a real
manual is lacking. The plan is for the actual documentation
to all live on the Logbook, with code-level comments only
living within the scripts.

Other changes need to be made now that the scripts are
imported into SVN: report $Id$ strings when appropriate,
etc.


git-svn-id: https://svn.hcro.org/mmm@668 9b194dde-2b4d-0410-98f0-bba1668eb47e