hex-proc-prep: last commit had big changes, and don't use autoref option
In the previous commit I accidentally committed changes to
hex-proc-prep to allow it to succeed as best it can in the condition
that all of the data for one of the polarizations are flagged. RAPID
likes to do this every so often.

The "autoref" option to newcalcal.csh is documented but is simply the
default behavior and isn't parsed, so don't pass it.



git-svn-id: https://svn.hcro.org/mmm@923 9b194dde-2b4d-0410-98f0-bba1668eb47e