hextoolkit: use os.path.walk instead of os.walk in buildsql
os.path.walk is better than os.walk because it calls a callback
progressively as it goes through the filesystem. os.walk, on
the other hand, builds up and returns a list of all of the files
and directories it's found. That list may end up being huge,
and os.walk's model requires two steps of building the list then
searching it, instead of a single pass through the filesystem.



git-svn-id: https://svn.hcro.org/mmm@978 9b194dde-2b4d-0410-98f0-bba1668eb47e