hex squint db: rename squint{x,y} to squint{az,el}
These names are more precise about the actual quantities that are
being computed, and are less likely to lead to confusion about
X and Y feeds. This may need to be revisited if we ever find it
worthwhile to do our gaussfitting in the RA/Dec coordinate system --
which would be easy to enable in hex-lib-gaussfit, but isn't
done since it doesn't seem well-motivated physically.



git-svn-id: https://svn.hcro.org/mmm@980 9b194dde-2b4d-0410-98f0-bba1668eb47e