[deepint] Enhance X axis labeling
Add support for displaying channels or frequency on the X axis.

Make X axis labeling of displayed channels smarter vis-a-vis user
linetype specification.

Change "mag_scale" keyword to "axis" keyword.  The "axis" keyword takes
two values: one for the xaxis and one for the yaxis.  The supported
xaxis values are "freq" (the default) and "chan".  The supported yaxis
values are "linear", "log", and "db".  Currently the default yaxis is
linear.

git-svn-id: https://svn.hcro.org/mmm@987 9b194dde-2b4d-0410-98f0-bba1668eb47e