[pgplotter] Make axes ranges more flexible
Now the ranges of axes an be specified as nil or a two element Array
[min, max] of Floats and/or nils.  If the axis range itself is nil, both
max and min are autoscaled.  If the axis range is a two element Array,
then any nil value in the Array means to autoscale that end of the axis
range.  This can be used, for example, to ensure that the lower end of
the Y axis is 0 while autoscaling the upper end like this...

plot(xx, yy, :yrange=>[0,nil])

git-svn-id: https://svn.hcro.org/mmm@1049 9b194dde-2b4d-0410-98f0-bba1668eb47e