[pgplotter] Allow autoscaling one end of range
Previously, the xrange, yrange, mag_range, and ph_range parameters only
supported autoscaling both ends of the range or neither end of the
range.  This patch allows autoscaling of only one end of a range.  Note
that in order to do this, the range must be passed as an Array rather
than a Range since Ranges cannot have a nil start or end.

git-svn-id: https://svn.hcro.org/mmm@1145 9b194dde-2b4d-0410-98f0-bba1668eb47e