[pgplotter] Provide thread safety
Calling Plotter#select from multiple threads could cause plotting from
one thread to interfere with another thread's plot and/or prevent proper
closing of Pgplot devices.

This change fixes that problem by allowing a block to be given to
Plotter#select.  The #select method will synchronize on an internal
Monitor object before calling pgslct.  If a block is given to #select,
it will be called (yielded to) before the Monitor is released, thereby
ensuring that all PGPLOT operations performed in the block will be
directed towards the proper device (i.e. plot).

All of the internal calls to #select have been converted to use this
block form, so most users should be unaware of the change (unless they
were already aware of the problems caused without this change :-)).

git-svn-id: https://svn.hcro.org/mmm@1194 9b194dde-2b4d-0410-98f0-bba1668eb47e