[uvfall] Use new (separate) color map for phase
The new color map for phase now wraps around at +/- 180 degrees so that
the colors appear continuous across phase wraps.

It also uses separate color maps for the phase "image" and the color bar
wedge.  The color map used for plotting includes black for "flagged"
data whereas the color map for the wedge includes just the phase colors.
This leads to a clear display of flagged data without affecting the
appearence of the color bar wedge.

git-svn-id: https://svn.hcro.org/mmm@1223 9b194dde-2b4d-0410-98f0-bba1668eb47e