ndshow.py: work-in-progress on interactive scaling setting
This is trying to emulate DS9 a bit. The current interaction
paradigm is TERRIBLE, but hopefully I can find a way to tweak
the algorithm a bit to make it a lot better. I think the
infrastructure for doing the interaction is fine; it's just
that the way that mouse gestures are translated into scaling
changes turns out to be really awkward.