Remove modules that I haven't used in a while, logging their existence.
I've been reluctant to delete some of the modules for fear that
I'll forget that I ever wrote them and end up duplicating the work.
To try to prevent that, I've logged them in "in-the-attic.txt"
(a la good old CVS/Attic) and hopefully next time I want to write a
new module, I'll check the attic first.