dualscal.py: add 'postinterval' option
This option sets the 'interval' item in the output gain dataset after
the calibration is computed, which controls how far MIRIAD is willing
to extrapolate gain solutions beyond the specified solution
interval. By default this value is set to the selfcal solution
interval, but you might want to apply the gains beyond those bounds if
the system parameters are stable on timescales longer than that value.
The 'postinterval' argument is measured in minutes.



git-svn-id: https://svn.hcro.org/mmm@1316 9b194dde-2b4d-0410-98f0-bba1668eb47e