pwmpfit.py: add a hacky routine for doing multiple LM iterations
I've been seeing a persistent problem where the LM algorithm exits
abnormally early: if you restart the algorithm with the solution that
it gives you, it returns you a nontrivially different, better
solution. I've been working on understanding this but haven't figured
out anything important so far; the algorithm as implemented here does
seem to match exactly with the original MINPACK implementation, which
one would expect to get everything right. So, in the meantime, at
least provide a tool for doing multiple iterations automatically.