pwmpfit/pwmpfit.py: rename variables and clean up fdjac2
Rename 'x0' to 'initial_params' where appropriate for clarity. Rename
'xall' to 'params' in fdjac2 since that what it is.

Change fdjac2 to only take the 'params' argument and set x =
params[ifree], since we have that array handy and the operation is
cheap.

In solve(), pass 'params' for the fdjac2 argument, instead of x0
(which became initial_params). I'm not sure why this didn't cause
more obvious problems but it seems like we were calling analytic
derivative functions with the wrong values for params: the initial
ones instead of the most recent ones.

In _manual_fdjac2(), use 'params' in self._ycall() instead of 'x'.
I believe this would have been buggy in any calls to _manual_fdjac2
with fixed parameters (not surprised if that never happened).