pwmpfit/pwmpfit.py: change integer status codes to a set of strings
The strings are (slightly) more comprehensible than integers, and
we can tell the caller all of the termination criteria that applied.
And we can invent new ones in the future. I also copied over the
documentation for the interpretations of the status codes, although
most of the important ones I find to be fairly unclear.