pwmpfit: fix that fucking bug
Certain fits of mine were exiting before true convergence was
achieved. After the past 16 hours of work, I've discovered an
off-by-one bug in _lmpar() that seems to cause the problem.  It's not
at all obvious to me how exactly it makes its effect, but I don't care
at this point. I've verified that the pwmpfit version and the official
Minpack fortran versions agree in their results. I've also verified
that moronsolve is no longer necessary.

Interesting that this bug wasn't caught in any of the lmder1 testcases
that I implemented. It obviously has something to do with
singularities in the matrix when finding the L-M parameter, so maybe
there's either something unusual in the n=1 case, or my particular
problem, which degenerates to y = ax, somehow triggers that case.