pwmpfit: add compatibility layer to use scipy.optimize.leastsq
Of course, a major major point of this module is to not have a
dependency on scipy.optimize, but it can be useful to be able to
cross-check results. Also, one might be interested in the nicer API of
pwmpfit but still want the driver to be the Fortran version. (Although
it lacks parameter limits, control over autoderivatives, etc, which is
another reason to be using the mpfit variant instead.)