scilib/srctable.py: sfind & deconv robustness improvements
The sfind docs say that you can sometimes get totflux < pkflux for
unresolved sources, simply due to it not enforing a minimum source
size. Enforce this ourselves, with a "pristine" keyword to disable
this (and any potential other) postprocessing.  Also force totflux_uc
to be positive.

In deconvolve(), add a default-on option to preserve the original
shape parameters in c_{major,minor,pa}.