scilib/astimage.py: correctly check FITS WCS axis units
We need to check the FITS WCS axis units to find out how to convert
them to radians. We do this by looking at the cunit[] values. However,
if there's no information about this in a FITS file, these values
start out unspecified. We need to call wcs.set() before checking the
units to get wcslib to fill in default values if applicable.