pwmpfit/pwmpfit.py: actually honor 'neg' derivative sidedness
Based on my reading of the code, negative sidedness wasn't being
honored unless there were also parameter upper limits. The old code
passed my little test case so I'm not sure what the deal was, but the
it certainly looked incorrect, and the new version still passes the
test case. And the change is straightforward enough that I'm too lazy
to look into this more deeply.