lmmin: renamed from pwmpfit
I think this name is better, and it doesn't appear to already be
taken. I'm guessing that Craig Markwardt chose "mpfit" to imply
"minpack curvefit", and neither of those items is particularly
relevant to the implementation at hand. (Minpack is certainly
important in terms of this module's provenance, but not in terms of
what it fundamentally does.)