intfbin/msimgen: use astutil, srctable, and quembed
Update to use my newer libraries, most importantly srctable.  Note
that we do not use astimage, because we try to set the image
units. That's an awfully small piece of functionality so it's probably
worth finding a workaround or implementing some dumb API for it in
astimage. When that conversion is ready, note that the fill_gauss
function has been reimplemented in intfbin/sfextract in a better way.