scilib/srctable.py: add a minimum axis product when deconvolving
If the deconvolved axis sizes are super small we should just treat
the source as a point. This is set to the pixel area in sfextract.
This was triggered by turning on the psfsize option in sfind, which
does seem to give better fits, but I think gives slightly off
results because of rounding in the sfind output.