lmmin/lmmin.py: prep for Great Transposition of Jacobians
The original MINPACK code is in Fortran, while we're using Python.
The two languages have opposite array index ordering conventions: in
Fortran, the first index variest fastest, while in Python, it's the
last. No accounting for this was made in the translation from Fortran
to Python, however. (I don't know the IDL convention.) This doesn't
affect the results, but it affects the speed and the cleanliness of
many of the constructs.

At some point I intend to go through and transpose all of the 2D
matrices in the Python code to get things in the proper order. This is
going to be a big undertaking. The only user-visible impact is on the
shape of the Jacobian matrix returned by the user Jacobian function,
if present.

This commit transposes the user-visible Jacobian without redoing
anything else just yet. This slows things down somewhat for now, but
means that when the Great Transposition finally happens, not all of
the user Jacobians will be broken. This will make life a lot easier
and will make me feel better about publicizing the code before the
G.T.