lmmin/lmmin.py: don't reallocate fjac continually
Allocate fjac once instead of reallocating it every time we want to
compute a Jacobian. This should be a bit more efficient.  We need to
be careful because for most of our purposes, we're ignoring the fixed
or tied parameters, but we need a full-sized fjac big enough to hold
all of the parameters since the user function doesn't know what's been
tied and what hasn't. Once the user function has been called the extra
rows are squeezed out.

This change also includes transposing the automatic Jacobian function,
which is actually very straightforward. In retrospect I'm not sure why
I thought it might be complicated.