scilib/astimage.py: fix up special axes and characteristic frequency
Upon further thought I realized that the "restfreq" stuff was totally
wrong; the definition of rest frequency is unrelated to what I'm
interested in, which is a characteristic frequency of the image.

The best way to get that value is from the spectral axis of the
coordinate system, but I was avoiding that because the spectral axis
might be velocity-based and there'd be a bunch of possible conversions
involved. But there's support in wcslib to do this stuff fairly
straightforwardly, so I exposed it in miriad-python (yay bundled copy
of wcslib) and added the right code to export 'charfreq', the desired
characteristic frequency. There might be some issues if some jerk uses
a coordinate system with weird units for the reference values. I want
to smack whoever decided that we needed configurable axis units.

In the process of all this it made sense to export a simple index for
the index of the spectral axis, a la wcsprm.spec, so I set that up and
converted _latlonaxes() to work the same way. This just makes sense
since it's a one-time computation that you either have the necessary
information for, or you don't.

Also update iminfo to dump the charfreq and mjd information.