scilib/srctable.py: give conservative and non-None uncerts if they're printed as zero
sfind prints uncertainties for source parameters with only a few decimal
places, and it's possible for them to get printed as all zeros even
though the fit succeeded. We used to convert these to None, but that
was making downstream code that the fit failed, since that condition
is often associated with uncertainties of asterisks, which we translate
to None. Instead, make a conservative assumption about the uncertainties.