intfbin/msphotom: add polarization processing
It turns out that doing this correctly requires a significant rework
of how we read in the data. But I think I'm doing it all more
correctly now.

In the process of implementing this I discovered a bug in casacore
related to the on-the-fly polarization processing. It doesn't actually
affect our results for these kinds of datasets, but I've submitted a
fix that I believe to be correct. And it could affect results if you
did anything with visibility weighting and/or Stokes conversions.