intflib/casaenv.py: emulate of casapy environment setup
Should allow us to script things with casa-python rather than
casa-script, which gets rid of the logfiles, annoying banners,
watchdogs, etc. I'm not sure if the plotting tasks will work
without all of the helper processes that casapy sets up, but
that's honestly kind of a low priority for me.