intflib/tasklib.py: selected CASA tasks as a library
I'm getting fed up with trying to trick CASA's crazy modules into
actually acting in a modular way. For much of the functionality, it
seems that the best approach is going to be to rewrite the Python
wrappers myself. The casautil module helps us do this in a relatively
painless way.

While we're at it, provide a program 'casatask' that lets these task
reimplementations be accessed from the command-line. This lets us
script CASA in the shell and gives another way for running the tasks
without the ridiculous overhead of the casapy environment.