intfmisc/casa-python.in: add hook for running scripts in a debugger
If you're trying to debug a casa-python script, you can't just prefix
your command-line with 'gdb --args' because casa-python is a shell
script that gdb doesn't know how to work with. But you can't just
run your script with 'gdb --args python ./script.py' because the
whole point of casa-python is that magic environment variables get
set up and a special Python interpreter has to be used for the
script to work.