intfbin/msptspec: new tool to give an idea of a source's spectrum from its visibilities
This is a clone of msphotom that bins things by spw instead of by time.
You could use it to measure a spectral index though to do that properly
you'd really want to do an MCMC kind of approach. It's too bad that there's
so much copy/paste between this and msphotom but I don't think there's
a way to really elegantly combine them.