intflib/tasklib.py: mfsclean: fix masks
We were creating and passing an empty mask which caused the imager to only
ever choose (0,0) for clean components. In my read of the task_clean code,
I thought the image was always created and passed to the imager, but it
seems to work to leave it empty if there's no masking.