intflib/tasklib.py: mfsclean: workaround for CASA's crazy coord parsing
If you give clean a phasecenter like 'J2000 1:2:3 4:5:6', it will not report
an error but it will give you ludicrously wrong coordinates, because it
wants 'J2000 1h2m3 4d5m6' instead. Try to catch this case and charge on ahead,
since a lot of my imaging is done in long batches, so I don't want to
have the task just abort if it can try to do the right thing. I marvel at
how the CASA parsing subsystem can give a grieviously wrong results without
reporting an error.