scilib/lsqmdl.py: vital doc/semantics fix
I was describing the weight parameter for the models as, well, a
weight when it should have been "inverse sigma", seeing as
you usually want to weight things by the inverse *variance*.
I believe this confusion also meant that the wrong weights were
being passed to polyfit() for the polynomial modeler. I'm not sure
how this snuck in since I've been careful about that kind of
terminology for a while.