Rework Bayesian Blocks analysis system
There were issues in the bayesian_blocks module from AstroML: some based on
mistakes in the original paper, some based on changes that I don't fully
understand -- in particular, there were no tstart and tstop parameters that
are needed to correctly blockify event series.

I also wasn't honoring the GTIs in the Chandra data, which was wrong and
caused problems in the high-ish rate case.