scilib/lsqmdl.py: improve PolynomialModel a bit
Rename 'degree' to 'nterms' since that makes it way easier for me
to infer the correct meaning of the parameter.

We were also sending np.polyfit degree=nterms when it *appears* we should have
been using degree=nterms-1. I feel like I must have been getting this right
before, so maybe this is related to the changes that popped up in Numpy 1.7?
But it seems very unlikely that they'd have broken compatibility like this,
either accidentally or intentionally.

Also compute rchisq manually, because the new polyfit function doesn't seem to
return the residual values correctly.