scilib/xbblocks.py: support multiple good-time intervals (GTIs)
As expected this required some thinking, but it wasn't too hard to implement
overall. We key off of the insight in Scargle+13, which is that gaps between
GTIs can just be ignored in the fundamental analysis.

xbin/chandrabblock was updated to match. It also got a few very small doc and
polish fixes.