scilib/lsqmdl.py: add MatMultComponent
This is for multiple independent components that are then mixed together via a
matrix multiply, where the multiplying matrix may itself be derived from an
indepent component.

I hope this turns out to be useful for more than just the problem I've been
working on, since it's taken a while to get it to work.