intfbin/casa-script: various fixes (see detailed message)
First, we try to compute the fully-qualified path of the input script and
always pass that to casapy. We move around in directories a bit and if casapy
can't open the `-c` argument as a file, it treats it as a line of Python,
leading to non-obvious failure modes.

Second, use some "trap" and "kill" magic to do our best to assassinate all of
the CASA subprocess if we get terminated. I think this code is overzealous
right now, and it's certainly not clean, but I'm pretty sure that it will do a
very robust job of making sure that CASA dies if we die.

Third, redirect CASA's input to /dev/null. This matters if the IPython shell
process dies, since by default it wants to read a line of input. This leads to
the script hanging forever in this case. (Yes, this happened.)

Fourth, hook up an environment variable that can cause the work directory to
be preserved, for debugging.