intflib/tasklib.py: avoid reconfiguring logger in some _cli() routines
Some of the tasks in this module have foo_cli() functions with no underlying
foo() implementation, because the _cli() API is as easy to use as the
underlying function would be. This is generally fine, but it was causing
inconsisent logging in my rxpackage processing scripts, because one thing the
_cli() functions *did* do that a library function wouldn't was reset the CASA
logger to be quieter. This meant that, if your processing invocation happened
to involve a step with a _cli() call, suddenly your logging would become a lot
less verbose for the rest of the processing steps. This suppresses information
and can make it look like steps aren't actually working, if you're not seeing
expected output.

I've added an "alter_logger" keyword argument to these functions so that the
processing script can signal that the logger configuration shouldn't be
changed.